Restructure code with action setup
This commit restructure the codebase to fit the name of the gem
and it also adds a basic GitHub action setup for this gem. This
also skips the current test for the moment and we will add more
test soon.
