+++ .github/workflows/ubuntu.yml
@@ -1,0 +1,30 @@
+name: ubuntu
+
+on:
+  push:
+    branches:
+      - master
+
+  pull_request:
+    branches:
+      - master
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - uses: actions/setup-ruby@v1
+        with:
+          ruby-version: 2.6.5
+
+      - name: Install bundler
+        run: gem install bundler
+
+      - name: Setup
+        run: bin/setup
+
+      - name: Run tests
+        run: bin/rspec
