--- .travis.yml
@@ -1,7 +1,0 @@
----
-sudo: false
-language: ruby
-cache: bundler
-rvm:
-  - 2.4.3
-before_install: gem install bundler -v 1.17.1
