+++ iev-termbase.gemspec
@@ -1,11 +1,10 @@
-
 lib = File.expand_path("../lib", __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "tc211/termbase/version"
+require "iev/termbase/version"
 
 Gem::Specification.new do |spec|
-  spec.name          = "tc211-termbase"
-  spec.version       = Tc211::Termbase::VERSION
+  spec.name          = "iev-termbase"
+  spec.version       = Iev::Termbase::VERSION
   spec.authors       = ["Ribose"]
   spec.email         = ["open.source@ribose.com"]
 
@@ -13,14 +12,11 @@ Gem::Specification.new do |spec|
   spec.description   = %q{Build scripts for the ISO/TC 211 Termbase}
   spec.homepage      = "https://open.ribose.com"
 
-  # Specify which files should be added to the gem when it is released.
-  # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
-  spec.files         = Dir.chdir(File.expand_path('..', __FILE__)) do
-    `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
-  end
-  spec.bindir        = "exe"
-  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
+  spec.bindir        = "exe"
+  spec.files         = `git ls-files`.split("\n")
+  spec.test_files    = `git ls-files -- {spec}/*`.split("\n")
+  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
 
   spec.add_runtime_dependency "creek"
   spec.add_runtime_dependency "relaton", "~> 0.5"
