+++ lib/iev/termbase.rb
@@ -1,0 +1,11 @@
+require "iev/termbase/version"
+
+module Iev
+  module Termbase
+    class Error < StandardError; end
+    # Your code goes here...
+  end
+end
+
+require "iev/termbase/term_workbook"
+require "iev/termbase/concept_collection"
