+++ lib/iev/termbase/cli/command.rb
