+++ lib/iev/termbase/concept.rb
