+++ lib/iev/termbase/concept_collection.rb
