+++ lib/iev/termbase/information_sheet.rb
