+++ lib/iev/termbase/version.rb
@@ -1,4 +1,4 @@
-module Tc211
+module Iev
   module Termbase
     VERSION = "0.1.12"
   end
