--- lib/tc211/termbase.rb
@@ -1,11 +1,0 @@
-require "tc211/termbase/version"
-
-module Tc211
-  module Termbase
-    class Error < StandardError; end
-    # Your code goes here...
-  end
-end
-
-require 'tc211/termbase/term_workbook'
-require 'tc211/termbase/concept_collection'
