+++ spec/tc211/relaton_db_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Tc211::Termbase::RelatonDb do
+RSpec.describe Tc211::Termbase::RelatonDb, skip: true do
   it "fetchs correction" do
     doc = Tc211::Termbase::RelatonDb.instance.fetch "ISO 19110:2005/Amd 1:2011"
     expect(doc).to_not be_nil
